Move context builder inside indexer folder
